Outlines an idea of merging files at $HOME
It is nice and all to have multiple castles as long as you don't try to
have the same files managed in multiple castles. I am, either, doing
something wrong as this is the first use case I have.

A machine at work and one at home, I need to manage my `.ssh/config`
properly and I don't like sharing keys among machines. Homesick (or its
alternatives) don't provide any work around this.

Home/Work machine setup is one thing, add Bastion or some
troubleshooting scripts to the mix and you have hell: either a dot file
repository that has every configuration and environment variable unneeded
for that environment or multiple castles that you always feel are missing
something that is in another castle.

My main idea is to have someone manage this for me. Homesick is pretty
transparent once the castle is loaded, allowing you to fuck things up.
Although I have no clue on how to tackle this, my feelings are that
hiding this management is worse: either this saves you time or it is
bullshit.

Carcassonne is a city-castle, it grew with the passing of years and saw
many different owners. A fitting name if this gets used by people and
not only me - as I am happy with the idea already.