Describes the use of suffixes to filter files
Some times, we don't want to merge all the files. We are already using
*prefixes* as the file name, so only *suffixes* are left for a pattern.
Previously, only OS names were specially used for suffixes so rather
than making it a "magic" use, turns it into a feature.

By explicitly defining that the `carcassone` command is not to be put
into the shell startup, we force this responsibility outside of the
command. Maybe into a shared castle, which may provide for many
different options to chose from.

TODO
----

Describe these ideas and give them a proper form:

* A naming convention for castles: `carcassone-*`
* Best practices for shared castles