Separates *program* from *standard* (with hooks)
Although this was pretty clear in my head, it needed (poor, lacking
enthusiasm) writing: the name Carcassonne can refer to two different
things, a *standard* or a *program*.

The *standard* is described, probably with a lot of flaws and omissions,
as a `dotfile` repository requiring a unique name and a "load script".
It should be rewritten in plain English for more people to understand,
being a quick read will probably increase the change of people
conforming to it.

The `help` output separates options between what they affect and all
possible uses of the program are now on display at "usage". An example
of how to use the program to merge files was added for clarity as well.

The hook system
----------------

Although only described in the *standard* without details, it always
seemed interesting to provide a way to auto-wire cloned repositories
(which now must be done manually or with another tool) into our home.
The best way to do this, now, is by calling `carcassonne` alone which
will reload some previous or peding configuration - or just run
everything again.

A hook will be a script cloned into "~.carcassonne" and executed
*before* or *after* the command is executed. Hooks will suffer from the
same sorting "merged files" suffer so we might benefit from a standard
priority in the future.

TODO
----

* Deprecate `CARCASSONNE_FILES` in favor of hooks.
* Study tools for testing using literal programming, hopefully this
  README can be trimmed down with them. Aurelio Jargas (verde) has
  something good, if I recall.