Describes hooks and simplifies explanation
The `load` hook is explicitly explained in a way I can say I though on
it before writing this, which is a lie I will always lie about.

Rearranges sections and tried to make explanations more direct, trying
to make a narrative following "what it does", "how it can be done
without it", "how it helps me do it". Hopefully things are clearer now.

I am almost feeling comfortable to start coding this. Maybe a good test
would be to try and simplify the README again, and see if the API
remains the same.

Almost forgot, sent this README to couple of people to take a look. I
don't expect them to understand anything (this is very confusing and
badly written) but questions might help prove myself an idiot or improve
the API.