Tests command under multiple shells
The idea of providing a single shell script is to make it portable and
useful on the majority of environments without many dependencies.
Continuous Integration is made for OSX and Linux with multiple shells
(bash, sh, busybox, ksh, tsh and zsh).

The tests are executed using *Literate Programming* so some
documentation is provided on how the program is structured and how we
expect it to work - along with automated testing.

For testing purposes, `-h` (help) and `-v` (version) are implemented.