Renames Carcassonne to Ore
Carcassonne is too big of a name and, as the following `sed`
substitutions shows, very error prone:

* s/carcassonne/ore/g
* s/carcarssonne/ore/g
* s/carcassone/ore/g

In case you want to reapply the `sed` command, you might need those:

* s/Carcassonne/Ore/g
* /github\.com\/topics/s/carcassone/castle-ore/
* /github\.com\/augustohp/s/carcassone/ore/

Along with realizing how much I like to repeat myself, some
little-last-hour improvements were made without breaking any
(un)existing behavior:

1. Makefile CARCASSONNE references exchanged by PROJECT
2. Makefile now calls variables and not the program directly
3. Makefile displays the `ln` and `$PATH` contents aiding debug
4. More exclusive GitHub topic `castle-ore`

Why `castle-ore` and not `ore` as GitHub topic
----------------------------------------------

Searching for compatible repositories is a *big feature* of the project
if adoption is a concern. "Carcassonne" is the name of a board game with
a bunch of projects around GitHub already, not much but the confusion
together with how difficult it is to write it made it easy to allow a
name change.

Using this project without the default behavior of "dotfile castles"
didn't make much sense as well. A install script would solve those
issues easily, without being compatible with how castles deal with
repetitive files - it makes sense to use that existing knowledge.

Carcassonne was a "city castle", analogous to merging castle files. I
though of "brick", "sand", "stone" and decided on "ore" for:

* It is the shortest word
* Has a "profit" meaning to it
* Is a mineral which can be stretched to the castle analogy :P

The sad part is that I don't know if I am happy with the name. I am
probably going to be unhappy if GitHub repositories start showing on the
topic search before I finish this.